Added Legal Material
I have Edited the Stock Creative Commons Zero v1.0 Universal (TEMPLATE), which is what this LICENSE.md file was based off of. And in the process added Some Personal Legal Portions Outlining the Use of My Property, by adding in a portion which sets the guidelines for what is Appropriate use, and what is not Appropriate of my Work.