Version 1.3 - minor fixes
fixed multiple issues

changelog:
- fixed a problem where if you clicked a button that did have the right
answer on it but it wasn't selected to have the correct answer on it, it
would be considered incorrect. this is presumably fixed
- fixed a problem where it would sometimes not change the values that
the UI buttons are already set to when they have a decimal value, or it
would generate answers with decimal values on problems that shouldn't
have decimal places such as subtraction problems. this has been fixed
for the most part.
- removed duplicate answers, the application will no longer generate
answers that are identical to the determined correct answer on any
problem